requirements.txt
braindecode==0.8.1
einops==0.7.0
matplotlib==3.8.0
mne==1.6.1
moabb==1.0.0
numpy==1.26.4
pyparsing==3.1.2
PyYAML==6.0.1
scikit_learn==1.3.1
seaborn==0.13.2
torch==2.2.1
torch_geometric==2.5.3
torcheeg==1.1.1
torchmetrics==1.3.1
torchvision==0.17.1
torchviz==0.0.2
tqdm==4.66.1
